requirements.txt
# Core ML libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0

# Deep Learning (uncomment if needed)
# torch>=1.9.0
# tensorflow>=2.6.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Data processing
openpyxl>=3.0.0
xlrd>=2.0.0

# Configuration
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Code quality
black>=21.0.0
flake8>=3.9.0
isort>=5.9.0

# MLOps and Experiment Tracking
mlflow>=1.28.0
wandb>=0.12.0
tensorboard>=2.6.0

# Model Serialization
joblib>=1.1.0

# API and Serving
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Monitoring and Observability
prometheus-client>=0.12.0
structlog>=21.1.0

# Database
psycopg2-binary>=2.9.7

# Utilities
tqdm>=4.62.0
click>=8.0.0
